refactor : chat 디렉토리 위치 변경 및 코드 수정
