+++ src/main/java/com/sparta/itsmine/domain/chat/dto/ChatRoomRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.itsmine.chat.dto;
+package com.sparta.itsmine.domain.chat.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
