+++ src/main/java/com/sparta/itsmine/domain/chat/dto/MessageType.java
@@ -1,4 +1,4 @@
-package com.sparta.itsmine.chat.dto;
+package com.sparta.itsmine.domain.chat.dto;
 
 public enum MessageType {
     // 메시지  타입 : 입장, 채팅
