+++ src/main/java/com/sparta/itsmine/domain/chat/dto/UserChatInfoResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.itsmine.chat.dto;
+package com.sparta.itsmine.domain.chat.dto;
 
 import lombok.Data;
