+++ src/main/java/com/sparta/itsmine/domain/chat/repository/BlackListRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.itsmine.chat.repository;
+package com.sparta.itsmine.domain.chat.repository;
 
-import com.sparta.itsmine.chat.entity.BlackList;
+import com.sparta.itsmine.domain.chat.entity.BlackList;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
