+++ src/main/java/com/sparta/itsmine/domain/chat/repository/ChatRoomAdapter.java
