+++ src/main/java/com/sparta/itsmine/domain/chat/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.itsmine.chat.repository;
+package com.sparta.itsmine.domain.chat.repository;
 
-import com.sparta.itsmine.chat.entity.ChatRoom;
+import com.sparta.itsmine.domain.chat.entity.ChatRoom;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
