+++ src/main/java/com/sparta/itsmine/domain/chat/repository/JoinChatAdapter.java
