+++ src/main/java/com/sparta/itsmine/global/common/response/ResponseCodeEnum.java
@@ -54,7 +54,7 @@ public enum ResponseCodeEnum {
     CHAT_GET_ROOM_LIST(HttpStatus.OK, "채팅방 리스트 입니다."),
     CHAT_SUCCESS_ROOM_LEAVE(HttpStatus.OK, "채팅방에서 나갔습니다"),
     CHAT_BLACKLIST_USER_ADD(HttpStatus.CREATED, "블랙 리스트 추가 했습니다."),
-    CHAT_BLACKLIST_USER_CANCEL(HttpStatus.CREATED, "블랙 리스트 취소 했습니다.");
+    CHAT_BLACKLIST_USER_CANCEL(HttpStatus.OK, "블랙 리스트 취소 했습니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
